fix: DynamicSequence/LazyDynamicSequence を統合し lower_bound の重大なバグを修正
両者は splay tree の骨格（zig/zag/access/split/split3/merge/merge3）が
完全に同一で、LazyDynamicSequence は lazy/has_lazy フィールドと apply()
を足しただけだった。F=void をデフォルトとする DynamicSequence<S, F=void>
に統合し、radix_heap/doubling と同じ void 正規化パターン
（std::conditional_t + [[no_unique_address]] + if constexpr/requires）
で1本化。

統合の過程で lower_bound() の重大なバグを発見: 直接降下版のアルゴリズムが
「ここまでの累積値」を全く追跡せず、各ノードの部分木全体の集約値を
毎回同じ key と比較していたため、答えが部分木の内部にある場合に誤った
結果を返していた（例: v=[1,1,1] で lower_bound(3) は 2 が正しいが 1 を
返す）。lib/test のどこからも呼ばれておらず未検証のまま埋もれていた
（旧 dynamic_sequence.hpp のみに存在し LazyDynamicSequence には
lower_bound 自体が無かった）。同じファイル内で正しく累積値を伝播している
max_right() を使って再実装することで修正し、乱数テスト（reverse との
組み合わせ含む）で naive 実装と一致することを確認済み。max_right/min_left
自体も乱数テストで独立に正しさを確認した。

AOJ 1508（DynamicSequence<Min<int>>）は実データ63ケース全て AC、
Library Checker の2問（lazy 版）もサンプルケースで AC を確認済み。
